test(sse): isolate ConcurrentWritesNeverCorruptFrames from -race
The splittingWriter introduced in PR #29's review fixup is
intentionally race-y by design — its `buf` field is unsynchronised
so two concurrent Write calls visibly interleave bytes, exposing
the bug the sse.mu prevents. Under -race, that intentional race
trips the data-race detector and the test fails for the design
reason, not because the system-under-test broke.

The local toolchain occasionally schedules the goroutines such
that no race is observed; CI runners hit it deterministically and
have started failing v0.7.2 PRs.

Same pattern as TestBashTimeout (PR #27): some tests deliberately
exercise concurrency in ways the race detector can't distinguish
from real bugs, and the cleanest fix is the conventional `!race`
build tag. The race detector itself catches the production case
(unsynchronised writes to s.w would race directly), so we don't
lose coverage by skipping this test under -race.

Verified:
  - go test ./internal/api/http/ -run TestSSEKeepalive
    → 4/4 pass (the splittingWriter test included)
  - go test -race ./internal/api/http/ -run TestSSEKeepalive
    → 3/3 pass (splittingWriter test excluded)
  - go test -race ./... → clean

Moved splittingWriter + the test into a new sse_norace_test.go
(`//go:build !race`) and dropped the now-unused runtime / providers
imports from the main sse_test.go.
