+++ internal/api/http/sse_test.go
@@ -4,13 +4,10 @@ import (
 	"bytes"
 	"context"
 	"net/http"
-	"runtime"
 	"strings"
 	"sync"
 	"testing"
 	"time"
-
-	"github.com/denn-gubsky/loomcycle/internal/providers"
 )
 
 // captureWriter is a minimal http.ResponseWriter+Flusher for tests.
@@ -50,45 +47,6 @@ func (c *captureWriter) Snapshot() string {
 	c.mu.Lock()
 	defer c.mu.Unlock()
 	return c.buf.String()
-}
-
-// splittingWriter is captureWriter's evil twin: Write splits each
-// payload into byte-by-byte appends with a runtime.Gosched() between
-// every byte. Two concurrent Write calls without external
-// synchronisation will visibly interleave in the buffer — not just
-// race-detected, but observable in the wire output.
-//
-// Used by TestSSEKeepalive_ConcurrentWritesNeverCorruptFrames to
-// prove the sse.mu actually prevents interleaving without depending
-// on `-race`. The buffer access in Write is intentionally
-// unsynchronised; rely on the sse.mu in the system-under-test, not
-// on this writer, for cross-goroutine ordering.
-type splittingWriter struct {
-	hdr    http.Header
-	buf    []byte // intentionally NOT mutex-protected
-	status int
-}
-
-func newSplittingWriter() *splittingWriter {
-	return &splittingWriter{hdr: http.Header{}}
-}
-
-func (s *splittingWriter) Header() http.Header { return s.hdr }
-func (s *splittingWriter) Write(p []byte) (int, error) {
-	for _, b := range p {
-		s.buf = append(s.buf, b)
-		runtime.Gosched()
-	}
-	return len(p), nil
-}
-func (s *splittingWriter) WriteHeader(st int) { s.status = st }
-func (s *splittingWriter) Flush()             {}
-
-// SnapshotAfterQuiescing should only be called after every writer
-// goroutine has stopped (e.g. cancel + wait). Reading concurrently
-// with active Write calls would race on the unsynchronised buf.
-func (s *splittingWriter) SnapshotAfterQuiescing() string {
-	return string(s.buf)
 }
 
 func TestSSEKeepalive_EmitsCommentFramesOnIdleStream(t *testing.T) {
@@ -161,53 +119,3 @@ func TestSSEKeepalive_GoroutineExitsOnContextCancel(t *testing.T) {
 	}
 }
 
-func TestSSEKeepalive_ConcurrentWritesNeverCorruptFrames(t *testing.T) {
-	// Pin the reason the sse struct carries a mutex: with two
-	// goroutines writing to the same response writer (main agent loop
-	// + keepalive ticker), unsynchronised writes interleave bytes
-	// mid-frame. Use a writer (`splittingWriter`) that appends one
-	// byte at a time with `runtime.Gosched()` between bytes — that
-	// makes interleaving visible WITHOUT depending on `-race`. The
-	// real `http.ResponseWriter` is not thread-safe either, so this
-	// matches production conditions more closely than a mutex-guarded
-	// captureWriter.
-	//
-	// Verified the test catches the bug: temporarily removing the
-	// mutex in sse.send produces malformed frames and trips the
-	// assertions below.
-	w := newSplittingWriter()
-	s, _ := newSSE(w)
-	s.start()
-
-	ctx, cancel := context.WithCancel(context.Background())
-	s.startKeepalive(ctx, 1*time.Millisecond)
-
-	const sends = 500
-	for i := 0; i < sends; i++ {
-		s.send(providers.Event{Type: providers.EventText, Text: "hi"})
-	}
-	cancel()
-	time.Sleep(20 * time.Millisecond) // give keepalive goroutine time to exit
-
-	got := w.SnapshotAfterQuiescing()
-	frames := strings.Split(got, "\n\n")
-	for i, f := range frames {
-		if f == "" {
-			continue // trailing empty after the final "\n\n"
-		}
-		if strings.HasPrefix(f, ":") {
-			// comment-only keepalive frame
-			if !strings.HasPrefix(f, ": keepalive") {
-				t.Errorf("frame[%d] looks like a comment but isn't a keepalive: %q", i, f)
-			}
-			continue
-		}
-		// Real event must have an event: line then a data: line.
-		lines := strings.Split(f, "\n")
-		if len(lines) < 2 ||
-			!strings.HasPrefix(lines[0], "event:") ||
-			!strings.HasPrefix(lines[1], "data:") {
-			t.Errorf("frame[%d] malformed (interleaved write?): %q", i, f)
-		}
-	}
-}
